Type Hardening, Consolidation, Assert Replacement
General:
chore: some code formatting
feat: changed guid definition to hoist the required type conversion, preventing the internal warning generation
feat: replaced macro std assert and macro assert usage with ::hg::assertd/rt
feat: fixed some template types by using underlying container type queries
feat: enforced typed constraints for type erase supported wrappers, preventing instantiation with incomplete types
feat: added explicit warning suppression to reduce warning noise when dealing with highly- and safely constrained type erase objects
feat: added overload struct helper to support type overloaded lambda expressions
feat: changed AutoArray<T, ...> to a concrete template type and added some utility function

+++ src/Engine.GFX.Render.Predefined/Effects/BrdfBase.cpp
@@ -375,7 +375,7 @@ static smr<AccelerationEffect> build_test_effect() {
 
 	/**/
 
-	Guid guid { "__Test__Proxy"_typeId.data, 0, 0, "BrdfBaseEffect"_typeId.data };
+	Guid guid { static_cast<u32>("__Test__Proxy"_typeId.data), 0, 0, "BrdfBaseEffect"_typeId.data };
 
 	return make_smr<AccelerationEffect>(
 		std::move(guid),
